requirements-dev.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.1.7
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-filter==2.4.0
django-multiselectfield==0.1.12
django-restframework==0.0.1
django-static-fontawesome==5.14.0.0
django-tempus-dominus==5.1.2.13
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.5.0
Markdown==3.3.4
Pillow==8.1.2
psycopg2==2.8.6
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
selenium==3.141.0
sqlparse==0.4.1
static3==0.7.0
stripe==2.57.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.4
zipp==3.4.1
